Handle field_names_prefix correctly for Bricks in FieldCollections
Before this, if set_field_names_prefix was called by a field group,
and a brick was added to the field group, the prefix was added to the
bricks fields instead of the name of the brick (which would then later
be applied to the field names).

Now, the prefix is applied to the bricks name which will the be applied
to the field names.